Adding data for testing. Changed quantity data type to float.

+++ databases/recipe_manager/recipe_manager.rb
@@ -30,7 +30,7 @@ def add_ingredients_to_recipe(recipe_id)
     name = gets.chomp
 
     print ">>Quantity: "
-    quantity = gets.to_i
+    quantity = gets.to_f
 
     unit_id = select_unit
